requirements.txt
-r wagtailbase/requirements.txt
-r ddhldap-django/requirements.txt
django-grappelli
django-haystack==3.1.1
django-model-utils
django-redis
django-require
elasticsearch>=5,<5.4
django-mysql
# mysql-python
mysqlclient
pillow
redis
south
django-autocomplete-light
simplejson
requests-oauthlib